Alauddin

+++ app/src/main/java/com/amtech/baseetcustomer/AddService/Model/ModelPlaceOrder.kt
@@ -3,5 +3,5 @@ package com.amtech.baseetcustomer.AddService.Model
 data class ModelPlaceOrder(
     val message: String,
     val order_id: Int,
-    val total_ammount: Int
+    val total_ammount: Double
 )
